[Cleanup] black, isort
