+++ graphein/grn/graphs.py
@@ -10,7 +10,7 @@ from typing import Callable, List, Optional
 import networkx as nx
 
 from graphein.grn.config import GRNGraphConfig
-from graphein.utils import (
+from graphein.utils.utils import (
     annotate_edge_metadata,
     annotate_graph_metadata,
     annotate_node_metadata,
