+++ graphein/utils/utils.py
@@ -9,7 +9,6 @@ from __future__ import annotations
 import os
 import sys
 from pathlib import Path
-from yaml import unsafe_load
 from typing import Any, Callable, Iterable, List
 
 import networkx as nx
@@ -17,6 +16,7 @@ import numpy as np
 import pandas as pd
 import xarray as xr
 from Bio.Data.IUPACData import protein_letters_3to1
+from yaml import unsafe_load
 
 
 def onek_encoding_unk(
@@ -34,7 +34,7 @@ def onek_encoding_unk(
 
 
 def filter_dataframe(df: pd.DataFrame, funcs: List[Callable]) -> pd.DataFrame:
-    """"""
+    """ """
     for func in funcs:
         df = func(df)
     return df
@@ -335,32 +335,3 @@ def import_message(
         print(f"    {installation}")
     else:
         print(f"{installation}")
-
-
-def parse_config(path: Path):
-    """
-    Parses a yaml configuration file into a config object
-    :param path: Path to configuration file
-    """
-    with open(path) as file:
-        config_dict = unsafe_load(file)
-
-    print(config_dict)
-    if config_dict["mode"] == "protein_graph":
-        return parse_protein_graph_config(config_dict)
-    elif config_dict["mode"] == "protein_mesh":
-        raise NotImplementedError
-    elif config_dict["mode"] == "rna":
-        raise NotImplementedError
-    elif config_dict["mode"] == "ppi":
-        raise NotImplementedError
-
-
-def parse_protein_graph_config(config_dict):
-    from graphein.protein.config import ProteinGraphConfig
-    config = ProteinGraphConfig(**config_dict)
-    print(config)
-    return config
-
-def parse_dssp_config(config_dict):
-    raise NotImplementedError
